#160 refactor: seperate TogglePostStatusModal
