+++ src/pages/recruitment-post/RecruitmentDetail.tsx
@@ -13,8 +13,7 @@ import EyeIcon from "@/components/icons/EyeIcon";
 import CommentIcon from "@/components/icons/CommentIcon";
 import BookmarkIcon from "@/components/icons/BookmarkIcon";
 import Button from "@/components/Button";
-import { Modal, ModalClose, ModalContent, ModalTrigger } from "@/components/Modal";
-import H3 from "@/components/text/H3";
+import TogglePostStatusModal from "@/components/TogglePostStatusModal";
 
 //더미 포스트 데이터 실제론 api 준비
 const dummyPostData: PostDetail = {
@@ -333,46 +332,5 @@ export default function RecruitmentDetail() {
         </div>
       </div>
     </div>
-  );
-}
-
-interface TogglePostStatusModalProps {
-  isClosed: boolean;
-  postId: string;
-}
-
-function TogglePostStatusModal({ isClosed, postId }: TogglePostStatusModalProps) {
-  const handleClick = () => {
-    //실제 api 연결
-    console.log(`${postId} 마감 상태 변경`);
-  };
-
-  return (
-    <Modal>
-      <ModalTrigger>
-        <Button color="error" variant="outline">
-          <Text>{isClosed ? "다시 열기" : "마감하기"}</Text>
-        </Button>
-      </ModalTrigger>
-      <ModalContent className="flex flex-col items-start justify-start p-2 space-y-5">
-        <H3 className="text-text-primary">
-          {isClosed ? "다시 열시겠습니까?" : "마감하시겠습니까?"}
-        </H3>
-        <div className="flex justify-end items-center w-full space-x-2">
-          <ModalClose>
-            <Button variant="outline" color="error">
-              <Text variant="subText" className="text-text-primary">
-                취소
-              </Text>
-            </Button>
-          </ModalClose>
-          <Button color="error" onClick={handleClick}>
-            <Text className="text-text-on-dark" variant="subText">
-              {isClosed ? "다시 열기" : "마감하기"}
-            </Text>{" "}
-          </Button>
-        </div>
-      </ModalContent>
-    </Modal>
   );
 }
